fix(fusion): MF-013 — widen vote_ones/vote_zeros to uint32_t
Sister to Finding 06. fused_bit_t.vote_ones/vote_zeros were uint8_t which
silently wraps at 256 revolutions. The fields are file-local — no header
exposes fused_bit_t — so widening is ABI-safe and isolated.

Same defense-in-depth as src/recovery/uft_sector_recovery.c (F06) and
src/recovery/uft_multiread_pipeline.c (F03+F06): vote counters track real
read counts, never bounded by data values.

+++ src/algorithms/advanced/uft_multi_rev_fusion.c
@@ -24,8 +24,8 @@ typedef struct {
     uint8_t value;              // 0 or 1
     float confidence;           // 0.0 - 1.0
     bool weak_bit;              // Inconsistent across revolutions
-    uint8_t vote_ones;          // How many revolutions had 1
-    uint8_t vote_zeros;         // How many revolutions had 0
+    uint32_t vote_ones;         // How many revolutions had 1 (>=256 supported)
+    uint32_t vote_zeros;        // How many revolutions had 0 (>=256 supported)
 } fused_bit_t;
 
 typedef struct {
